Broken widgets on create invoice
