+++ lib/routes/create_invoice/widgets/animation_progress.dart
@@ -1,0 +1,19 @@
+import 'dart:math';
+
+class AnimationProgress {
+  final Duration duration;
+  final Duration startTime;
+
+  const AnimationProgress({
+    required this.duration,
+    required this.startTime,
+  });
+
+  double progress(Duration time) => max(
+    0.0,
+    min(
+      (time - startTime).inMilliseconds / duration.inMilliseconds,
+      1.0,
+    ),
+  );
+}
