+++ lib/routes/create_invoice/widgets/animation_properties.dart
@@ -1,0 +1,6 @@
+enum AnimationProperties {
+  X,
+  Y,
+  COLOR1,
+  COLOR2,
+}
